+++ internal/service/kubernetes/utils.go
@@ -3,7 +3,9 @@ package kubernetes
 import (
 	"encoding/base64"
 
+	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
 
 func setB64Decoded(d *schema.ResourceData, field string, value string) error {
@@ -13,4 +15,17 @@ func setB64Decoded(d *schema.ResourceData, field string, value string) error {
 	}
 
 	return d.Set(field, string(decoded))
+}
+
+func storageEncryptionSchema(description string, computed bool) *schema.Schema {
+	return &schema.Schema{
+		Description: description,
+		Type:        schema.TypeString,
+		Computed:    computed,
+		Optional:    true,
+		ForceNew:    true,
+		ValidateDiagFunc: validation.ToDiagFunc(validation.StringInSlice([]string{
+			string(upcloud.StorageEncryptionDataAtReset),
+		}, false)),
+	}
 }