+++ CHANGELOG.md
@@ -5,6 +5,11 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+- kubernetes: support for node group custom plans
+- kubernetes: support for data at rest encryption in node groups
+
 ## [5.3.0] - 2024-05-13
 
 ### Added